+++ v8/protocol/src/org/chromium/sdk/internal/v8native/protocol/input/data/PropertyWithRef.java
@@ -12,11 +12,11 @@ public interface PropertyWithRef {
   @JsonOptionalField
   void value();
 
-  long ref();
+  int ref();
 
   @JsonOptionalField
-  long attributes();
+  int attributes();
 
   @JsonOptionalField
-  long propertyType();
+  int propertyType();
 }
